split for pre-1.4 testing
